Convert to csproj (#10)
* Migration to csproj

* trying to get creating nuget packages working again

* Remove old proj files

* trying to get scripts embeded

* Getting more precision for ms time

* add ticks

* bump to 0.0.8

* fix csproj files to use version instead of versionprefix for now

+++ src/Zoxive.HttpLoadTesting.Client/Domain/Iteration/Dtos/IterationDto.cs
@@ -15,9 +15,9 @@ namespace Zoxive.HttpLoadTesting.Client.Domain.Iteration.Dtos
 
         public int Iteration { get; set; }
 
-        public int StartTick { get; set; }
+        public long StartTick { get; set; }
 
-        public int EndTick { get; set; }
+        public long EndTick { get; set; }
 
         public long UserDelay { get; set; }
